masleads/pyproject.toml
[tool.poetry]
name = "masleads"
version = "0.1.0"
description = "MasLeads Backend Developer Challenge"
authors = ["Armando Ezequiel Puerta <armando.ezequiel.puerta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
flask-restx = "^1.1.0"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
psycopg2-binary = "^2.9.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
